move test from cli to create
